Create extension methods to support key types of `PedometerStepKind`
