+++ lib/src/winrt/foundation/collections/imap.dart
@@ -16,6 +16,7 @@ import '../../../types.dart';
 import '../../../utils.dart';
 import '../../../winrt_helpers.dart';
 import '../../devices/sensors/enums.g.dart';
+import '../../devices/sensors/pedometerstepkind_helpers.dart';
 import '../../internal/ipropertyvalue_helper.dart';
 import '../../internal/map_helpers.dart';
 import '../../media/mediaproperties/mediapropertyset.dart';
@@ -147,7 +148,7 @@ class IMap<K, V> extends IInspectable
     }
 
     if (isSameType<K, PedometerStepKind>()) {
-      return _lookup_PedometerStepKind_COMObject(key as PedometerStepKind);
+      return lookupByPedometerStepKind(key as PedometerStepKind) as V;
     }
 
     if (isSameType<K, String>()) {
@@ -242,26 +243,6 @@ class IMap<K, V> extends IInspectable
     if (FAILED(hr)) throw WindowsException(hr);
 
     return IPropertyValue.fromRawPointer(retValuePtr).value;
-  }
-
-  V _lookup_PedometerStepKind_COMObject(PedometerStepKind key) {
-    final retValuePtr = calloc<COMObject>();
-
-    final hr =
-        ptr.ref.lpVtbl.value
-                .elementAt(6)
-                .cast<
-                    Pointer<
-                        NativeFunction<
-                            HRESULT Function(
-                                Pointer, Int32, Pointer<COMObject>)>>>()
-                .value
-                .asFunction<int Function(Pointer, int, Pointer<COMObject>)>()(
-            ptr.ref.lpVtbl, key.value, retValuePtr);
-
-    if (FAILED(hr)) throw WindowsException(hr);
-
-    return _creator!(retValuePtr);
   }
 
   V _lookup_String_COMObject(String key) {
@@ -396,7 +377,7 @@ class IMap<K, V> extends IInspectable
     if (isSameType<K, GUID>()) return _hasKey_GUID(value as GUID);
     if (isSameType<K, int>()) return _hasKey_Uint32(value as int);
     if (isSameType<K, PedometerStepKind>()) {
-      return _hasKey_PedometerStepKind(value as PedometerStepKind);
+      return hasKeyByPedometerStepKind(value as PedometerStepKind);
     }
     if (isSameType<K, String>()) return _hasKey_String(value as String);
 
@@ -470,28 +451,6 @@ class IMap<K, V> extends IInspectable
     }
   }
 
-  bool _hasKey_PedometerStepKind(PedometerStepKind value) {
-    final retValuePtr = calloc<Bool>();
-
-    try {
-      final hr = ptr.ref.lpVtbl.value
-              .elementAt(8)
-              .cast<
-                  Pointer<
-                      NativeFunction<
-                          HRESULT Function(Pointer, Int32, Pointer<Bool>)>>>()
-              .value
-              .asFunction<int Function(Pointer, int, Pointer<Bool>)>()(
-          ptr.ref.lpVtbl, value.value, retValuePtr);
-
-      if (FAILED(hr)) throw WindowsException(hr);
-
-      return retValuePtr.value;
-    } finally {
-      free(retValuePtr);
-    }
-  }
-
   bool _hasKey_String(String value) {
     final retValuePtr = calloc<Bool>();
     final hValue = convertToHString(value);
@@ -552,7 +511,7 @@ class IMap<K, V> extends IInspectable
     }
 
     if (isSameType<K, PedometerStepKind>()) {
-      return _insert_PedometerStepKind_COMObject(
+      return insertByPedometerStepKind(
           key as PedometerStepKind, (value as IInspectable).ptr);
     }
 
@@ -648,31 +607,6 @@ class IMap<K, V> extends IInspectable
     }
   }
 
-  bool _insert_PedometerStepKind_COMObject(
-      PedometerStepKind key, Pointer<COMObject> value) {
-    final retValuePtr = calloc<Bool>();
-
-    try {
-      final hr = ptr.ref.lpVtbl.value
-              .elementAt(10)
-              .cast<
-                  Pointer<
-                      NativeFunction<
-                          HRESULT Function(
-                              Pointer, Int32, COMObject, Pointer<Bool>)>>>()
-              .value
-              .asFunction<
-                  int Function(Pointer, int, COMObject, Pointer<Bool>)>()(
-          ptr.ref.lpVtbl, key.value, value.ref, retValuePtr);
-
-      if (FAILED(hr)) throw WindowsException(hr);
-
-      return retValuePtr.value;
-    } finally {
-      free(retValuePtr);
-    }
-  }
-
   bool _insert_String_enum(String key, WinRTEnum? value) {
     final retValuePtr = calloc<Bool>();
     final hKey = convertToHString(key);
@@ -756,7 +690,7 @@ class IMap<K, V> extends IInspectable
     if (isSameType<K, GUID>()) return _remove_GUID(key as GUID);
     if (isSameType<K, int>()) return _remove_Uint32(key as int);
     if (isSameType<K, PedometerStepKind>()) {
-      return _remove_PedometerStepKind(key as PedometerStepKind);
+      return removeByPedometerStepKind(key as PedometerStepKind);
     }
     if (isSameType<K, String>()) return _remove_String(key as String);
 
@@ -790,16 +724,6 @@ class IMap<K, V> extends IInspectable
         .value
         .asFunction<
             int Function(Pointer, COMObject)>()(ptr.ref.lpVtbl, key.ptr.ref);
-
-    if (FAILED(hr)) throw WindowsException(hr);
-  }
-
-  void _remove_PedometerStepKind(PedometerStepKind key) {
-    final hr = ptr.ref.lpVtbl.value
-        .elementAt(11)
-        .cast<Pointer<NativeFunction<HRESULT Function(Pointer, Int32)>>>()
-        .value
-        .asFunction<int Function(Pointer, int)>()(ptr.ref.lpVtbl, key.value);
 
     if (FAILED(hr)) throw WindowsException(hr);
   }
